Add modules for different sections of the page
