+++ modules/date.js
@@ -1,0 +1,3 @@
+import { DateTime } from 'luxon';
+
+const date = new DateTime();
